refactor(pages/race): move race view component to seperate folder
