+++ apps/client/src/pages/race/templates/race-view/RaceTrack.stories.tsx
