+++ apps/client/src/pages/race/templates/race-view/race-background/RaceBackground.template.tsx
@@ -1,6 +1,6 @@
 import { ReactElement, useEffect, useRef } from 'react';
 
-import spritesheet from '../../../../assets/spritesheets/city-tiles-map.png';
+import spritesheet from '../../../../../assets/spritesheets/city-tiles-map.png';
 import {
   getRaceTrackHeight,
   getRaceTrackNRows,
