+++ apps/client/src/pages/race/templates/race-view/race-background/draw-blocks.ts
